Cache optimization on hostname

+++ src/main/java/org/cloudfoundry/promregator/scanner/ReactiveAppInstanceScanner.java
@@ -178,40 +178,47 @@ public class ReactiveAppInstanceScanner implements AppInstanceScanner {
 	
 	private Mono<String> getApplicationUrl(String applicationId, String protocol) {
 		String key = applicationId;
-
-		Mono<RouteEntity> routeMono = this.cfAccessor.retrieveRouteMapping(applicationId)
-			.map(rm -> rm.getResources())
-			.map(l -> l.get(0))
-			.map(e -> e.getEntity())
-			.map(entity -> entity.getRouteId())
-			.flatMap(routeId -> { 
-				return this.cfAccessor.retrieveRoute(routeId); 
-			})
-			.map(routeResponse -> routeResponse.getEntity());
-		// WARNING! routeResponse.getApplicationsUrl() is the URL back to the application
-		// and not the URL which points to the endpoint of the cell!
-
-		Mono<String> domainMono = routeMono.map(route -> route.getDomainId())
-			.flatMap(domainId -> {
-				return this.getDomain(domainId);
-			});
 		
-		Mono<String> applicationUrlMono = Mono.zip(domainMono, routeMono)
-			.map(tuple -> {
-				String domain = tuple.getT1();
-				RouteEntity route = tuple.getT2();
-				
-				String url = String.format("%s://%s.%s", protocol, route.getHost(), domain);
-				if (route.getPath() != null) {
-					url += "/"+route.getPath();
-				}
-				
-				return url;
-			});
-		
-		this.hostnameMap.put(key, applicationUrlMono);
-		
-		return applicationUrlMono;
+		synchronized (key.intern()) {
+			Mono<String> applicationUrlMono = this.hostnameMap.get(key);
+			if (applicationUrlMono != null) {
+				return applicationUrlMono;
+			}
+			
+			Mono<RouteEntity> routeMono = this.cfAccessor.retrieveRouteMapping(applicationId)
+				.map(rm -> rm.getResources())
+				.map(l -> l.get(0))
+				.map(e -> e.getEntity())
+				.map(entity -> entity.getRouteId())
+				.flatMap(routeId -> { 
+					return this.cfAccessor.retrieveRoute(routeId); 
+				})
+				.map(routeResponse -> routeResponse.getEntity());
+			// WARNING! routeResponse.getApplicationsUrl() is the URL back to the application
+			// and not the URL which points to the endpoint of the cell!
+	
+			Mono<String> domainMono = routeMono.map(route -> route.getDomainId())
+				.flatMap(domainId -> {
+					return this.getDomain(domainId);
+				});
+			
+			applicationUrlMono = Mono.zip(domainMono, routeMono)
+				.map(tuple -> {
+					String domain = tuple.getT1();
+					RouteEntity route = tuple.getT2();
+					
+					String url = String.format("%s://%s.%s", protocol, route.getHost(), domain);
+					if (route.getPath() != null) {
+						url += "/"+route.getPath();
+					}
+					
+					return url;
+				});
+			
+			this.hostnameMap.put(key, applicationUrlMono);
+			
+			return applicationUrlMono;
+		}
 	}
 	
 	private Mono<String> getDomain(String domainIdString) {
